requirements.txt
async-timeout==4.0.2
cachelib==0.9.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
Flask-Caching==2.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
pytest==7.1.3
python-decouple==3.6      
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
